requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
azure-storage-blob>=12.20
azure-identity>=1.16
python-dotenv>=1.0
